LayoutTests/fast/frames/sandboxed-iframe-history-denied-expected.txt

CONSOLE MESSAGE: line 1: Sandbox access violation: Blocked a frame at "null" from accessing a frame at "".  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
CONSOLE MESSAGE: line 1: TypeError: undefined is not an object (evaluating 'top.history.back')
ALERT: PASS
This test verifies that a sandboxed IFrame cannot navigate the top-level frame using the history API.
